fix: bug for really big diffs
