--- src/lib/git/detect_staged_changes.ts
@@ -1,8 +1,0 @@
-import { gpExecSync } from '../utils/exec_sync';
-export function detectStagedChanges(): boolean {
-  return (
-    gpExecSync({
-      command: `git --no-pager diff --no-ext-diff --cached`,
-    }).length > 0
-  );
-}
